Don't hang when logging a failed pdflatex run
Added the -halt-on-error arg when running pdflatex so that it
doesn't wait for input on error and just dies.

The command was also changed to not use shell features (&&) and instead
just run it twice. I changed this because the docs say: "If shell is
True, it is recommended to pass args as a string rather than as a
sequence." but having cmd as a list works much better.