lots of tidying and refactoring
Each section of the course is now represented by a subclass of Item.
This is in charge of working out the destination path and URL of the
item, its markdown representation, and so on.
The course structure consists of a forest of Item objects.

There was a tonne of repetition in the pandoc code - now there's just
one method, which takes an item, a template file, and a format (html or
pdf)

Content is now piped in and out of pandoc, removing the need for temp
markdown files. It would be nice to pipe into plastex too, though it
will always produce extra files for things like tikz images.