restructuring; use pathlib
This commit restructures some of the build process, so it'll be more
easily extensible. It also uses pathlib throughout, because it's a
neater way of handling file paths.

Rather than referring to the `config` dictionary directly, there are methods
on the MakeCourse object to get things like the build directory and the web
root.

Use context blocks to read files:

    with open(path) as f:
        content = f.read()

The builder class is constructed from a variable list of superclasses in
`main`. Ideally, things like decktape should be optional and only
included when needed.

`copy_theme_static_files` has moved to the Theme object: if there's ever
a different kind of theme, it could override that method.

`PlastexRunner.getEmbeddedImages` was repeating a lot of code for
markdown images and tikz pictures; now there's a list of regular
expressions and just one method. Rather than doing a global replace of
the path, it only replaces the exact instance of the path it found.