use a visitor pattern for the processing step
Running pandoc to make HTML and building PDFs are two independent
steps. There was also a lot of duplicated code to deal with a structure
that uses parts and one that doesn't.

Now, we first check whether the structure uses parts, and stop if
there's a mix of parts and chapters at the top level, then run a
sequence of 'processors' on the structure. A processor is a visitor
pattern which visites each of the items in the course structure, and
either does something or silently passes over it.