remove pandoc when unnecessary; themes use jinja2
Themes use jinja2 to render HTML output. Each item has an `as_html`
method which produces HTML for the item's content.

makecourse can look for themes in a hierarchy of directories:
    * config['themes_dir']
    * course_directory/themes
    * makeCourse/themes
    * ./themes

This is so that you don't have to clone all the themes into each course,
and the package can provide a default theme.

Pandoc is only used to create PDFs (though that's broken at the moment)

The Python Markdown package converts markdown to HTML, and PlasTeX takes
LaTeX straight to HTML without going through markdown.

Slides and PDFs are broken at the moment.

I'd like the ability to override PlasTeX templates in the theme.

I *think* I've finally got a handle on URLs for items and images.