move custom plasTeX renderer here; fix logging; plastex code is in a subpackage
It felt iffy having the custom plasTeX renderer used by this project in
a fork of plasTeX. Now that we're using plasTeX as a Python package
rather than a command-line tool, the renderer can be part of the
makeCourse package.

This commit adds some trickery so that plasTeX doesn't break the logging
for everything else.

The \numbas, \youtube etc macros were added onto plasTeX.Packages.embed,
which again means using a fork of plasTeX.
Instead, they're now defined in makeCourse.plastex.macros.