add makecourse.sty
There's a lot of boilerplate repeated at the top of TeX files. It's best
in a package, so it can be provided by makeCourse and the author just
needs to do \usepackage{makecourse}.

I haven't looked at many courses to see what could go in this package;
at the moment it's just got the embed macros.