Create splitter reveal.js plugin
The markdown reveal.js plugin can be used to convert markdown text
to reveal.js slides on the fly. It contains a wonderful feature where
it can split up slides based on a data-separator attribute in the html
source.

The feature is extremely useful, but the markdown parsing that the
plugin was designed for is not, because coursebuilder already runs the
content through its own markdown compiler and produces raw HTML.
We also can't directly use the marked.js compiler by dumping the source
into the reveal.js slides because it is not flexible enough for us.

This would normally not matter so much, because I could just pass
the markdown reveal.js plugin raw HTML straight from makecourse,
but the markdown plugin also escapes characters, e.g. changes ampersands
to &amp; This breaks mathjax.

So, I have taken the data-separator part of the markdown.js plugin,
but removed the bits where it runs things through marked.js. The result
is a reveal.js plugin where the slides are still split based on a separator
in the way I would like, but without any further processing of the text content.

This should fix mathjax in slides.