The -l option makes relative links
The 'local file paths' mode used to just stick the absolute path of the
build directory on the front of absolute URLs. That doesn't work when
serving through a local HTTP server.

Instead, it now works out the right number of '../' to put in front of
the URL, based on the location of the item being rendered.