Render tikz images with new pdf2svg renderer
This commit takes some committed but as yet unreleased changes to
plasTeX and implements them as part of the makeCourse plasTeX
overrides. Specfically, the pdf2svg imager is added, and the tikz
packages have been re-implemented to directly use that imager.

Once a new version of plasTeX incorporating these changes has been
released, these additions can be removed and we can rely on the built in
tikz package and simply set the 'vector-imager' option to 'pdf2svg'.