Switch default imager from dvipng to pdftoppm
When producing DVI output latex doesn't support reading the size of
PDF images, throwing the error "Cannot determine size of graphic".
This creates problems with HTML output when including PDF graphics
because the default non-vector imager is dvipng.

This commit changes the default imager to pdftoppm, which uses the
pdflatex compiler instead with PDF as the intermediate format, solving
the problem.

This change requires an extra prerequiste to be installed so that the
pdftoppm command is available.