Support alt text on graphicx and tikz images
This commit adds support for a new \alt{} command that can be used in
the same way as \label{} or \caption{} to add alt text to an image.
The command populates an altText property on the last image or tikzfigure
inserted into the document.

When the altText property exists on an image it is parsed and output
to the HTML version of the document. The PDF version omits the alt
text, justified by being primarily a print format.