Add custom argType for graphicx package
A custom argType is now loaded when importing the graphicx package
to help with parsing includegraphics arguments. The new custom argType
is almost identical to a standard plastex dictionary argument type
but explicitly checks for the keys "width" and "height" and parses
them as dimensions. The other arguments are parsed in the standard way.

This helps because a standard pastex dictionary can only have a single
subtype, but the includegraphics arguments could be either numbers
(scale), dimensions (width/height) or boolean (keepaspectratio). Using
this custom dict argument type better handles widths and heights of
the form 0.5\texwidth because they are then correctly calculated as
dimensions.