Revert "copy a 'static' directory in the same dir as an item, if it exists"
This reverts commit 2675348. With that commit multiple copies of the
static directory are placed throughout the build output tree. This is
not good because it can create many copies of large files such as videos,
which are then not even used because the path filtering rewrites files
to point to the static directory that happens to be under the theme
name, e.g. [build]/default/static.