Allow images with relative path in markdown source
This commit adds a markdownRenderer extension that detects images
referenced relative to the markdown source file and copies the source
image into the build tree, where required. This is closer to how the
latex renderer handles images.

With this, images no longer have to be stored in the global course
'static' directory and referenced in markdown source relative to
'/static'.

Additional benefits are that referenced images are stored under the
content output directory rather than higher up the tree and also that
any markdown previews will likely work better when the source has image
paths relative to the file.