Check if .tex was recently built and don't rebuild
This uses the -recorder feature of pdflatex to produce a '.fls' file
during processing. The file contains a list of all files read as input
and output.

If a .fls file exists, makecourse will read it and check the modified
time on all the input files to see if it can skip rebuilding the .tex
source.

Since (by default) the .fls file is cleaned up with other latex aux
files after processing, this only works if makecourse is run in "lazy"
mode, with the -z flag.