Handle verbatim definitions inside \newenvironment
This changes the verbatim environment so that verbatim/comment can be
used inside \newenvironment{}{}, like in LaTeX when the verbatim
package is loaded.

A new test is added to the verbatim environment to search for a
\end{...} that matches the surrounding environment. If found, the
\end{...} is inspected to see if it really contains an (e.g.)
\endverbatim macro. If it does the \end{...} is expanded and output,
otherwise the verbatim environment continues as normal.

This should fix situations such as:

\newenvironment{answer}{\comment}{\endcomment}
Question \begin{answer}Answer\end{answer}