rejig the numbas embed filter
The HTML that replaces <numbas-embed> is now in a template file
belonging to the theme.

The JS code linking Numbas embeds to the feedback display is a bit
tidier.

The Numbas iframe isn't present in the produced HTML - it's inserted
when shown, so that the iframe isn't collapsed when it loads.

CSS classes are used to hide/show bits of the feedback display, instead
of setting the `display` property.