Cache document structure and reload it when valid
When processing document item types a document structure is generated
by running plastex and looking at the results. The document structure is
then used later while rendering output.

With this commit, the induced structure is cached to a .dst file and
if the cache remains valid it is loaded and used on the next run. Now
plastex does not need to be run every time as long as the cache is
valid, i.e. the document has not changed.

A document structure cache is considered valid if the usual rules of
source and output file modification times are satisfied, and if the
splitlevel of the document (according to config.yml) has not been changed.