Switch mathjax to CHTML and include mathtools
This should hopefully help with some rendering problems in the PDF
version of markdown content when using the SVG renderer.

The mathtools package was also requested a while ago. It now looks to be
included as part of mathjax and so has been included in the list of
packages to load.