Split pdf output for document item types
For document item types, plastex splits the output into multiple html
files based on the splitlevel set on the item (default 0, indicating to
split on chapters).

This commit adds logic to look at the output of pdflatex and split the
output .pdf file into multiple files to match as best as possible the
plastex split output.

This works by parsing the .toc file output by pdflatex for ToC entries
corresponding to the chapters/sections plastex has split on. The .pdf
file is also parsed (using PyPDF2) to find Link Objects, which allows us
to find the real page numbers the splits should occur on (rather than
just page number labels, which can be roman or even reset).

For this to work well, the output from pdflatex must include a ToC with
Link objects, and ToC entries for each point the pdf file is to be split
at.