Set a default root_url if none in config.yml
This does not matter much, it is only used if chirun is asked to
produce absolute paths to content, using the root_url template.

It's also used in the Chirun LTI Provider, but only as a way to
work out the path to send students to, which changes depending on
if a year and/or theme directory is included in the path.