add a blank comment to html passed to HTMLFilter before it's passed to bs4
fixes #134

bs4.BeautifulSoup produces a warning when it thinks you're passing in a
filename. The assumption is that you'll only use it to parse full
documents, not fragments.

So one solution would be to prepend `<!doctype html>`, but that makes it
add in html, body, and p tags, which would be a faff to remove.

Prepending a single comment avoids the warning, and doesn't seem to
cause anything else to be added, so the comment node can easily be
removed after parsing.