abstract out alternative formats for items; extract headers from HTML
There are now potentially three alternative formats for an item (slides;
PDF; Notebook), so it makes sense for the Item object to produce an
abstract list of alternative formats instead of repeating a load of code
for each one.

Also, the HTMLFilter extracts headers from the HTML, to produce a table
of contents, instead of doing this in JS once the page has loaded.