splitting markdown into sections happens in the markdown renderer
The old theme rendered markdown to HTML, then put it in a textarea to be
handled by reveal.js's markdown renderer, which would also split frames
across the `---` marker.

This commit changes the markdown renderer so that the content is split
into sections on `---`, and each section is rendered separately, then
they're concatenated.

This removes the mdx_outline extension. It just seemed to be used to
invent <section> tags based on headers, but that's not necessary.