Improvements to image handling
see #167

This changes the plastex renderer and imagers to intercept PDF files
used as the source for `\includegraphics`. They're immediately converted
to SVG using `pdf2svg`, instead of going into the fallback LaTeX
document, which is slow and error-prone.

Need to look at how to set the size of the resulting image in HTML.