use webR 0.2.1
This uses an official release of webR, loaded from webr.r-wasm.org.

This fixes #161 and #165.

The string rendering of the result of a block of R code doesn't match
the R repl. It should be rewritten to match eventually, but I don't have
it in me to try to do that now!