implement \hfil using flexbox
see #84

This relies on browser support for the :has() CSS function.

Elements who have an `\hfil` command as an immediate child are forced to
have `display: flex`, and the `\hfil` has `flex-grow: 1`.

I haven't tested this with anything other than a minimal example so I
don't know how robust it is.

I'm quite unsure about starting to implement the TeX spacing commands:
the layout model is completely different, so it's unlikely we'll really
satisfy anyone's demands to get the same spacing effects in both PDF and
HTML output.