use @counter-style for list item symbols
fixes #171

Chrome doesn't support the symbols() function for anonymous list counter
styles, so each list has to use the @counter-style at-rule in a <style>
tag to make up the list of symbols.

This isn't supported by Safari in the current public release, but is
supported by v17 technology preview so will start working in Safari
soon. For now, Safari shows the default markers for <ol> and <ul>
elements.