hide icon defs; slides body gets overflow: hidden
fixes #191

Chrome was displaying an extra scrollbar because the recently-introduced
SVG of icon definitions was displayed at the bottom of the page, under
the <main> tag.

This commit both hides the icon defs, because they shouldn't have been
visible anyway; and sets `overflow: hidden` on the <body> tag in slides,
so that if anything else makes it overflow then there won't be an extra
scrollbar.