PDF build: load files from a temporary HTTP server
see #186

In order to avoid some web security restrictions, the headless browser
must load pages from an http:// address instead of file://.

This commit changes the `renderer.get_browser` function to become a
context manager which starts an HTTP server on an OS-assigned port. The
server terminates when the context ends.

I've also taken the opportunity to abstract out the repeated code
shared by the base renderer and the slides renderer.