slides: handle font size better; tighten spacing
I noticed that on very narrow viewports, the text was too big so slides
overflowed.

For slides view, I think it's a reasonable principle to ensure the
viewport is always a certain number of characters wide.
This commit changes the font size from
`max(3vmin, 24px)`
to
`min(3vw,4vh)`.

This means that text can end up smaller than 24px, and a box in 4:3
aspect ratio always fits inside the viewport.

Comparing the HTML and PDF versions of the beamer slides in the sample
course, I decided to halve the vertical spacing around paragraphs,
headers, and display maths.