don't show hidden items in the structure; add a hash to URLS for hidden items
fixes #197

The old LTI tool used the "is_hidden" property of structure items to
control access, as part of the PHP script.

The new LTI tool serves package output as static files, and we'd like a
solution for packages served outside the LTI tool too.

This commit changes the templates for the structure pages to not include
items that are hidden.

These items are still built, so a hash is added to their URLs so that
students can't guess them.

The hash uses the item's slug and a salt string, either given in the
course config's `hash_salt` property, or passed as a command-line
argument `--hash-salt`. The LTI tool will use the package's editing
UUID as the salt.