reduce the sidebar break point; sidebar can shrink
See #184

The breakpoint for putting the sidebar above the text is now at 55em.

When to the side of the text, the sidebar will take up 30% of the
available width or the length of the longest line, whichever is shortest.
This means it usually shrinks a bit before the breakpoint is hit.