Build a zip of the output, and link to it from the intro page.
When the `build_zip` config option is true, a .zip file is created in
the root of the build directory, containing all of the output files
except for directories corresponding to hidden items.
There's a link to the zip from the introduction page.
The zip's filename is based on the slug of the package's title. It's
recorded in the manifest, under the key `zipfile`.

fixes #207