don't use document title in plastex filename rules
fixes #214

Sometimes the document has a title which is an empty
string.

This happened when a single .tex file had two document environments.

plasTeX can't cope with it and throws an IndexError when picking a
filename for the output.

Just not using the title in the filename's wildcard spec avoids this.