fix thmenv and proof templates
When rendering an environment, the begin and end macros each produce an
element with the name of the environment, so they use the same template.
You can tell which end you've got by looking at the obj.macroMode
attribute.

This commit changes the templates for thmenv and proof to render just
the appropriate half of the wrapper.