thinking about how to resolve refs in math mode
see #219

This adds an overridden handler for the \ref command in MathJax.

plasTeX is supposed to add a <script> tag with the id `plastex-labels`
to the end of the document, giving a mapping from LaTeX label names to
their text presentations.

The MathJax ref handler looks up labels in this dictionary and makes
reference nodes as appropriate.

This works for single-page `chapter` items, but `document` items need
more thought: plasTeX splits up the HTML into several files, and the
`plastex-labels` script is only included on the top-level file.

It looks like there's also a problem with text-mode refs in this case,
anyway: the URLs from plasTeX, which are relative to its output path,
aren't rewritten by Chirun.