add plastex-labels dictionary to each file produced by plasTeX
The plasTeX renderer's cleanup method can post-process files before
they're written.
This commit adds a post-processing function which uses the course's
`make_relative_url` method to produce a list of URLs for the labels
defined in the document, relative to that section, which might be an
extracted section from a bigger document.

I also took the time to improve `make_relative_url` so it doesn't
include unnecessary `..` parts.

The course `force_theme` property seems not to be used any more - it was
only set by the `url` template filter when called with a theme argument,
which isn't done any more: each theme is built independently of the
others.