stub beamer's \only macro
the macro `\only<n>{content}` is supposed to show its contents only on
the given frame number, when a frame is rendered several times.

This commit adds a stub implementation of `\only`, which consumes the
number and renders the content. Once we've worked out how to do
multi-part frames, the logic about when to include the contents can be
implemented.