add links to the previous and next items in the top nav
fixes #203

The top nav shows links to the previous and/or next item, just under the
breadcrumbs.

Items which are hidden, or have `pager: false` in their config, are
skipped over.

In order to do this, document items have to have their structure
expanded before they're rendered, so that has moved to a separate
process, `ExpandDocumentProcess`.