Improve LaTeX PDF splitting
Rather than a regex, plasTeX is now used to parse the .aux file.
This helps it cope with part entries, which don't use a \numberline
command in the TOC entry.

I'm not entirely happy that the splitting is robust: we need to test if
it misses out initial pages that aren't part of the structure that comes
from the TOC.

Because the PDF doesn't record the span of pages for a particular part
of the document, just the first page, the code assumes that sections
don't overlap.
I don't know how to work around this; could we check the position on
which the next part starts?

The split PDFs are all disjoint. This isn't good: when there are several
levels, I'd expect the PDF for Part A to contain all of the pages of
Chapter AA, for example.
So the routine for finding end pages should look for the next item *of
the same level or higher*, not just the next item.