improvements to icons in dark mode
The 'invert image colours' filter only applies to images inside the item
content, not icons in the Chirun interface.
This fixes #246

The link colour is now #ccf instead of #ffe.

The icons use `var(--bg-colour)` where they previously used white.
This means that in dark mode, the PDF icon is a sheet of black paper
with a light outline, for example.