LaTeX PDFs: if the title is not manually set, get it from the source
If the title for an item with LaTeX source is not explicitly set in the
config, then it is set using either the title from the resulting PDF or
using the stem of the source file name.

The title is only set in the PDF if you use the `\hypersetup` command,
e.g.

```
\hypersetup{
    title={My document}
}
```

Just using `\title` isn't enough.

fixes #272