Add keyboard interactivity to the sidebar table of contents
The table of contents in the sidebar is enriched by JS which adds
keyboard interactivity.

Only one link in the TOC is in the tab order, and then you can use arrow
keys to move between items, and expand or collapse lower levels.

The Home and End keys move to the first and last visible items.

There's type-ahead search: with the table of contents focused, typing
letters will move focus to the first item matching the typed word.

Items in the TOC whose corresponding sections are currently visible are
highlighted.

fixes #276