try to make includegraphics insensitive to the case of the file's extension
A package had a file `graphic.JPG`. The command
`\includegraphics{graphic}` didn't find it, because it was only looking
for `graphic.jpg`.

This changes the `\includegraphics` command to glob all files whose
names start with the given name, and then return the first file whose
extension, converted to lower case, is in the expected list.