Packages can override theme templates with files in a 'templates' folder.
This needs more careful thought and testing before merging in.
It's likely that anyone doing this would want to do it across lots of
packages.

This changes the jinja template loader to try `<course root>/templates`
before `<theme template path>`.

Template paths prefixed with "base/" always use the theme's template, so
that a template provided by the package can extend the base template.

For example, in `<package>/templates/item.html`:

    {% extends "base/item.html" %}

    ...

The location of the package's `templates` directory should be
configurable, like the `static` directory.

It's possible that we'll just build up lots of ways to break backwards
compatibility like we did in Numbas, when packages routinely extend or
redefine things that we'd like to change.