add a "Custom CSS code" display option
This adds a "custom CSS code" field to the display options. The code is appended to the page. This could be used to add user-specific changes that it doesn't make sense to make universal.

Examples include colours for headings; loading a new font; more finely controlling font weight.

This shouldn't be used for changes that are likely to be useful to lots of people: I'd like to continue improving the standard display options.

related to #146