switch to the narrow-viewport layout when text scale is >= 200%
The narrow-viewport layout used a media selector, which isn't affected by a change in the --font-scale CSS variable.

This adds a bit to the customiser's `update_page` method to add a class `.big-text` to the <body> element when the font scale is 200% or more.
When that happens, the page layout is switched to have the sidebar above the main content, like on a narrow viewport.